+++ src/forze_postgres/dependencies/__init__.py
@@ -1,5 +1,0 @@
-from .module import postgres_document_configurable, postgres_module, postgres_txmanager
-
-# ----------------------- #
-
-__all__ = ["postgres_module", "postgres_document_configurable", "postgres_txmanager"]