+++ src/forze_s3/dependencies/__init__.py
@@ -1,5 +1,0 @@
-from .module import s3_module, s3_storage
-
-# ----------------------- #
-
-__all__ = ["s3_module", "s3_storage"]